+++ packages/graphiql-toolkit/src/create-fetcher/lib.ts
@@ -1,7 +1,6 @@
 import { DocumentNode, visit, GraphQLError } from 'graphql';
 import { meros } from 'meros';
 import { createClient, Client, ClientOptions } from 'graphql-ws';
-import { SubscriptionClient } from 'subscriptions-transport-ws';
 import {
   isAsyncIterable,
   makeAsyncIterableIteratorFromSink,
@@ -116,9 +115,16 @@ export const createWebsocketsFetcherFromClient = (wsClient: Client) => (
     }),
   );
 
-export const createLegacyWebsocketsFetcher = (
-  legacyWsClient: SubscriptionClient,
-) => (graphQLParams: FetcherParams) => {
+/**
+ * Allow legacy websockets protocol client, but no definitions for it,
+ * as the library is deprecated and has security issues
+ *
+ * @param legacyWsClient
+ * @returns
+ */
+export const createLegacyWebsocketsFetcher = (legacyWsClient: {
+  request: (params: FetcherParams) => unknown;
+}) => (graphQLParams: FetcherParams) => {
   const observable = legacyWsClient.request(graphQLParams);
   return makeAsyncIterableIteratorFromSink<FetcherResult>(
     // @ts-ignore
@@ -181,13 +187,14 @@ export const getWsFetcher = (options: CreateFetcherOptions) => {
   if (options.wsClient) {
     return createWebsocketsFetcherFromClient(options.wsClient);
   }
-  if (options.legacyClient) {
-    return createLegacyWebsocketsFetcher(options.legacyClient);
-  }
   if (options.subscriptionUrl) {
     return createWebsocketsFetcherFromUrl(
       options.subscriptionUrl,
       options.wsConnectionParams,
     );
+  }
+  const legacyWebsocketsClient = options.legacyClient || options.legacyWsClient;
+  if (legacyWebsocketsClient) {
+    return createLegacyWebsocketsFetcher(legacyWebsocketsClient);
   }
 };